till all products

+++ app/src/main/java/com/sacoding/prs/data/models/Articles.kt
@@ -1,5 +1,5 @@
 package com.sacoding.prs.data.models
 
 data class Articles(
-    val Articles_id: List<Double>
+    val Articles_id: List<Long>
 )
